Remove notification class
