+++ app/src/main/java/io/github/wulkanowy/WulkanowyApp.kt
@@ -16,7 +16,7 @@ import io.fabric.sdk.android.Fabric
 import io.github.wulkanowy.BuildConfig.DEBUG
 import io.github.wulkanowy.data.repositories.preferences.PreferencesRepository
 import io.github.wulkanowy.di.DaggerAppComponent
-import io.github.wulkanowy.services.sync.factory.MainWorkerFactory
+import io.github.wulkanowy.services.factory.MainWorkerFactory
 import io.github.wulkanowy.utils.CrashlyticsTree
 import io.github.wulkanowy.utils.DebugLogTree
 import timber.log.Timber
