+++ app/src/main/java/io/github/wulkanowy/services/sync/BaseNotification.kt
@@ -1,0 +1,9 @@
+package io.github.wulkanowy.services.sync
+
+import android.content.Context
+import androidx.core.app.NotificationManagerCompat
+
+class BaseNotification(protected val appContext: Context) {
+
+    protected val notifacationManager by lazy { NotificationManagerCompat.from(appContext) }
+}
