+++ app/src/main/java/io/github/wulkanowy/services/sync/SyncManager.kt
@@ -1,0 +1,8 @@
+package io.github.wulkanowy.services.sync
+
+class SyncManager {
+
+    fun initialize() {
+    }
+}
+
