+++ app/src/main/java/io/github/wulkanowy/services/sync/grade/GradeNotification.kt
@@ -1,0 +1,5 @@
+package io.github.wulkanowy.services.sync.grade
+
+class GradeNotification {
+}
+
